+++ Requirements/Php/websharks-php-phar-building-requirement.rb
@@ -2,7 +2,7 @@ require File.join(File.dirname(__FILE__), 'websharks-php-requirement')
 
 class WebsharksPhpPharBuildingRequirement < WebsharksPhpRequirement
   def satisfied?
-    `php -i| grep phar.readonly`.downcase.include? "off"
+    `php -i | grep phar.readonly`.downcase.include? "off"
   end
 
   def message